Add 'romfile' to nic definition
QEMU nics have built-in ROM files used during early BIOS boot to
enable PXE booting.  One can specify a path to a different version
via the 'romfile' parameter of the a nic definition.  Optionally,
the string 'off', which is defined in qcli v0.0.9 as
qcli.DisabledNetDeviceROMFile will result in passing the parameter
'romfile=' making it empty, which prevents the built-in ROM loading,
removing it as a boot device.

- bump qcli to v0.0.9 for romfile disable feature
- update vm-network doc with disable example