feat: add support for configuring/injecting cloud-init into VMs
Update to qcli v0.3.1 which added support for a vvfatt driver enabling
machine to create and attach a directory as a VFAT disk in the guest.

Add a new CloudConfig portion to the API that can import and use
user-data, meta-data and network-config.  Definiting these values in
VM's config will render out the defined user-data, network-config and
meta-data (auto generated if not supplied). This device is detected as
as NoCloud data source and will initilized the guest if cloud-init is
present in the image.

Additional changes:

- moved utility functions to pkg/api/utils.go
- added unittests for cloud-init API
- Added 'test' and 'test-api' make targets
- Added example configuration with cloud-init
- Updated README to point to examples directory
- Fixes some missing arguments to prints/logs found by go test
- Add 'make test' to github actions

Signed-off-by: Ryan Harper <rharper@woxford.com>